Let fileProvider be more strict in serving files
url.parse sometimes results in misleading results while parsing request.url,

Here an example of the result of url.parse("http://www.google.com"):

{ protocol: 'http:',
  slashes: true,
  host: 'www.google.com',
  hostname: 'www.google.com',
  href: 'http://www.google.com/',
  pathname: '/' }

As the server is only listening to variable pathname,
it may think that the client is interested in our "/" instead of
"/" on an other location (as in this case for google.com).

Since the fileProvider should only take care about serving local files,
be sure that each request.url starts with a "/".