Make sure that file provider reports errors
Note:

There is currently a strange bug, where the request handler is closed
before the file provider is actually requesting the request handler to close.

As consequence of this problem, the request handler creates a log before the
file provider is updating the information. If for example a not readable file
was causing errors, it would log the 200 OK status code into the log
and send the 500 Server Error status code to the browser.

I am very curious what part of the mechanisme is creating such a behavior.